+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Cache logic.
 - Find Transaction status logic.
+- Example data for swagger.
 
 ### Changed